fix: add missing file ProductFilterType enum and update response DTOs for improved data handling
- Add ProductFilterType enum to categorize products as ALL, BEST_SELLER, TRENDING, HIGH_RATED, and DISCOUNTED.
- Change averageRating to double, and update reviewCount and saleCount to long in ProductListResponse for better precision and scalability.
- Remove products field from PromotionDetailResponse to streamline the response structure.

+++ shopme-webparent/shopme-client/src/main/java/com/shopme/client/dto/response/ProductListResponse.java
@@ -18,7 +18,7 @@ public class ProductListResponse {
     private float discountPercent;
     private int discountPrice;
 
-    private float averageRating;
-    private int reviewCount;
-    private int saleCount;
+    private double averageRating;
+    private long reviewCount;
+    private long saleCount;
 }
